backend/requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.2
Pillow==10.0.0
platformdirs==3.10.0
psycopg==3.1.10
psycopg2==2.9.7
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.24.3
